requirements.txt
biopython==1.81
click==8.1.3
numpy==1.24.3
pandas==1.5.3
scipy==1.10.1
simpletransformers==0.63.9
torch==2.0.0